[master] Cache docker agian

+++ project.sh
@@ -92,8 +92,10 @@ case $1 in
   run $node_image 'yarn install --no-bin-links --cwd=compiler'
 
   echo "Clone QMK with submodules."
-  git clone -b zkf --single-branch --recurse-submodules https://github.com/zored/qmk_firmware.git $QMK_DIR || ls $QMK_DIR/quantum
-  rm -rf $QMK_DIR/.git
+  if [[ ! -f "$QMK_DIR/Makefile" ]]; then
+    git clone -b zkf --single-branch --recurse-submodules https://github.com/zored/qmk_firmware.git $QMK_DIR
+    rm -rf $QMK_DIR/.git
+  fi
 
   echo "Checking firmware flasher presence..."
   if [[ $wally != '' ]] && [[ ! -e $wally ]]; then
